Completed the new JobPostings view
